♻️ refactor: 자주 사용되는 함수 분리리
